+++ apps/rlm_web/lib/rlm_web_web/controllers/page_controller.ex
@@ -1,0 +1,7 @@
+defmodule RlmWebWeb.PageController do
+  use RlmWebWeb, :controller
+
+  def home(conn, _params) do
+    render(conn, :home)
+  end
+end